% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff_template.R
\name{diff_template}
\alias{diff_template}
\title{Compare an edited report template to the standard report template}
\usage{
diff_template(analysis, file_path)
}
\arguments{
\item{analysis}{A character string specifing the desired analysis template. Current options are: "tsp".}

\item{file_path}{A character string specifying the path to the edited template file.}
}
\value{
String with the file path for the html rendering of the diff. This can be passed to
"submit_diff" if the user chooses to anonymously post their diff results to the web for easier sharing.
}
\description{
Compare an edited report template to the standard report template
}
\details{
This function uses diffr and htmlwidgets to render a file difference
between an existing package template and a user-edited template. An HTML
file is rendered and saved which displays the highlighted differences in
the two files. The function will create a pop-up for the user to choose where to
save the output - PLEASE name this "[filename].html" for whatever filename you choose.
}

