% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicate_template.R
\name{duplicate_template}
\alias{duplicate_template}
\title{Duplicate a report template}
\usage{
duplicate_template(analysis)
}
\arguments{
\item{analysis}{A character string specifing the type of analysis desired. Current options are: "tsp".}
}
\value{
The file path to the duplicated file, should the user later desire to
supply it to "diff_template".
}
\description{
Duplicate a report template
}
\details{
This function duplicates an existing report template should the
user decide to make edits to any portion of the standard template. The function
will pop-up a file chooser dialog box so that the user may select a directory
in which to create the duplicated file. PLEASE name your file "[filename].Rmd" - the system will
create an empty file for you and copy the contents of the orignal template into it.
We encourage the user to use the "diff_report" command if they make changes to the original report.
}

